Feature: Add saved progress checkpoints.
1. Seasons that have not been completed, timeout
from rate limiting, or are exited early from
`Ctrl+C` save its progress in a checkpoint pickle
file and attempt to resume this progress at a
later time.

2. Rather than waiting on an exponential timeout
increase, the progress is saved and the script
will attempt a request from another team if the
maximum attempts have been tried and failed.